# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Fitness platform REST API — training programs, nutrition diary, Google Calendar integration.",
        "title": "FitProfi API",
        "contact": {},
        "version": "0.1.0"
    },
    "host": "localhost:8086",
    "basePath": "/api/v1",
    "paths": {
...
        "/integrations/google/connect": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Redirects the authenticated user to Google OAuth consent screen.",
                "tags": [
                    "Integrations"
                ],
                "summary": "Connect Google Calendar",
                "responses": {
                    "302": {
                        "description": "Redirect to Google OAuth"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/nutrition/entries/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns a single diary entry by ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nutrition"
                ],
                "summary": "Get nutrition diary entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.DiaryEntry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Replaces the meal type and food items of an existing diary entry.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nutrition"
                ],
                "summary": "Update nutrition diary entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Updated entry data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_controller_nutrition.UpdateEntryRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.DiaryEntry"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Deletes a diary entry belonging to the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Nutrition"
                ],
                "summary": "Delete nutrition diary entry",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "Entry ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/profiles/user": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the fitness profile of the authenticated user.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Get user profile",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.UserProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Replaces the fitness profile for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Update user profile",
                "parameters": [
                    {
                        "description": "Updated profile data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_controller_profiles.UpdateUserProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.UserProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates a fitness profile for the authenticated user.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Profiles"
                ],
                "summary": "Create user profile",
                "parameters": [
                    {
                        "description": "User profile data",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_controller_profiles.CreateUserProfileRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.UserProfile"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/login": {
            "post": {
                "description": "Authenticates the user and returns a JWT token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Login",
                "parameters": [
                    {
                        "description": "Login credentials",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_controller_authorize.LoginRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_application_command_authorize.LoginUserResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
...
        "/users/reset-password": {
            "patch": {
                "description": "Resets the user password using a one-time reset token.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Auth"
                ],
                "summary": "Reset password",
                "parameters": [
                    {
                        "description": "Reset token and new password",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/internal_delivery_controller_authorize.ResetPasswordRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "github_com_msskobelina_fit-profi_internal_application_command_authorize.LoginUserResult": {
            "type": "object",
            "properties": {
                "email": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "token": {
                    "type": "string"
                },
                "userID": {
                    "type": "integer"
                }
            }
        },
...
        "github_com_msskobelina_fit-profi_internal_delivery_controller.ErrorResponse": {
            "type": "object",
            "properties": {
                "error": {
                    "type": "string",
                    "example": "validation error message"
                }
            }
        },
...
        "github_com_msskobelina_fit-profi_internal_domain_model.DiaryEntry": {
            "type": "object",
            "properties": {
                "createdAt": {
                    "type": "string"
                },
                "date": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "items": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.DiaryItem"
                    }
                },
                "mealType": {
                    "type": "string"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                }
            }
        },
        "github_com_msskobelina_fit-profi_internal_domain_model.DiaryItem": {
            "type": "object",
            "properties": {
                "calories": {
                    "type": "number"
                },
                "carbsG": {
                    "type": "number"
                },
                "createdAt": {
                    "type": "string"
                },
                "entryId": {
                    "type": "integer"
                },
                "fatG": {
                    "type": "number"
                },
                "grams": {
                    "type": "number"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "proteinG": {
                    "type": "number"
                },
                "updatedAt": {
                    "type": "string"
                }
            }
        },
...
        "github_com_msskobelina_fit-profi_internal_domain_model.Goal": {
            "type": "string",
            "enum": [
                "lose_weight",
                "gain_weight",
                "rehab",
                "keep_fit",
                "competition"
            ],
            "x-enum-varnames": [
                "GoalLoseWeight",
                "GoalGainWeight",
                "GoalRehab",
                "GoalKeepFit",
                "GoalCompetition"
            ]
        },
...
        "github_com_msskobelina_fit-profi_internal_domain_model.UserProfile": {
            "type": "object",
            "properties": {
                "age": {
                    "type": "integer"
                },
                "createdAt": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "fullName": {
                    "type": "string"
                },
                "goal": {
                    "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.Goal"
                },
                "id": {
                    "type": "integer"
                },
                "updatedAt": {
                    "type": "string"
                },
                "userId": {
                    "type": "integer"
                },
                "weightKg": {
                    "type": "number"
                }
            }
        },
...
        "internal_delivery_controller_authorize.LoginRequest": {
            "type": "object",
            "required": [
                "email",
                "password"
            ],
            "properties": {
                "email": {
                    "type": "string",
                    "example": "john@example.com"
                },
                "password": {
                    "type": "string",
                    "example": "secret123"
                }
            }
        },
...
        "internal_delivery_controller_authorize.ResetPasswordRequest": {
            "type": "object",
            "required": [
                "password",
                "token"
            ],
            "properties": {
                "password": {
                    "type": "string",
                    "minLength": 6,
                    "example": "newSecret123"
                },
                "token": {
                    "type": "string",
                    "example": "eyJhbGciOiJIUzI1NiJ9..."
                }
            }
        },
...
        "internal_delivery_controller_nutrition.UpdateEntryRequest": {
            "type": "object",
            "required": [
                "items",
                "mealType"
            ],
            "properties": {
                "items": {
                    "type": "array",
                    "minItems": 1,
                    "items": {
                        "$ref": "#/definitions/github_com_msskobelina_fit-profi_internal_domain_model.DiaryItem"
                    }
                },
                "mealType": {
                    "type": "string",
                    "enum": [
                        "breakfast",
                        "lunch",
                        "dinner",
                        "snack"
                    ],
                    "example": "dinner"
                }
            }
        },
...
        "internal_delivery_controller_profiles.CreateUserProfileRequest": {
            "type": "object",
            "required": [
                "age",
                "fullName",
                "goal",
                "weightKg"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 25
                },
                "description": {
                    "type": "string",
                    "example": "I want to stay healthy"
                },
                "fullName": {
                    "type": "string",
                    "example": "John Doe"
                },
                "goal": {
                    "type": "string",
                    "enum": [
                        "lose_weight",
                        "gain_weight",
                        "rehab",
                        "keep_fit",
                        "competition"
                    ],
                    "example": "keep_fit"
                },
                "weightKg": {
                    "type": "number",
                    "example": 75.5
                }
            }
        },
...
        "internal_delivery_controller_profiles.UpdateUserProfileRequest": {
            "type": "object",
            "required": [
                "age",
                "fullName",
                "goal",
                "weightKg"
            ],
            "properties": {
                "age": {
                    "type": "integer",
                    "example": 26
                },
                "description": {
                    "type": "string",
                    "example": "Updated description"
                },
                "fullName": {
                    "type": "string",
                    "example": "John Doe"
                },
                "goal": {
                    "type": "string",
                    "enum": [
                        "lose_weight",
                        "gain_weight",
                        "rehab",
                        "keep_fit",
                        "competition"
                    ],
                    "example": "lose_weight"
                },
                "weightKg": {
                    "type": "number",
                    "example": 74
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Enter: Bearer \u003ctoken\u003e",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}